fix: use the native youtube view for the native
It's opening the in app
tracked at
https://github.com/sarbagyastha/youtube_player_flutter/issues/462
