+++ client/pubspec.yaml
@@ -25,6 +25,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
+  flutter_inappwebview: ^5.3.2
   grpc: ^3.0.0
   intl: ^0.17.0
   protobuf: ^2.0.0
